Updated Home.md with Relevant Information
The team name and logo were added to the Home file, along with the students' names and ID numbers of the group members. The information from the template is still there, but as a hidden comment within the file that you can access by editing the Home.md file. Once the file is finalised, all of the irrelevant comments can be deleted.