Updated index.md
A basic description was added of what the game might involve to provide an example and idea of what is being required in that section. The start date was added for when we officially start working on the project, and the end date is for when we finish and hand the project over. A few milestones were added, and the client's name and dates were added to the signatures section at the bottom of the file.